Generalized logo component
