+++ packages/components/src/Logo/Logo.Shapes.tsx
@@ -1,0 +1,52 @@
+import * as React from "react"
+
+export default {
+  Contiamo: {
+    svg: (
+      <svg viewBox="0 0 1740 360">
+        <path d="M93.869,283.579c-31.861,-31.93 -49.275,-78.675 -41.663,-126.448c9.095,-57.083 54.741,-106.57 112.674,-119.197c10.342,-2.254 20.923,-3.352 31.549,-3.335c38.192,0.192 76.071,16.199 103.329,42.946c0.389,0.27 0.607,0.582 0.889,0.883c1.778,1.896 1.982,4.717 -0.104,6.948l0,0.001l-0.001,0.001l-0.006,0.006l-51.919,51.919c-0.862,0.809 -1.22,0.952 -1.885,1.184c-1.42,0.497 -2.913,0.132 -4.109,-0.719c-0.762,-0.372 -1.496,-0.879 -2.156,-1.504c-19.836,-18.528 -52.364,-21.787 -75.812,-6.922c-13.025,8.258 -22.867,21.534 -26.696,36.609c-7.372,29.017 9.762,62.436 39.857,72.714c21.527,7.352 46.688,2.23 63.659,-14.203c0.901,-1.042 2.27,-1.721 3.959,-1.721c0.104,0.006 0.209,0.013 0.313,0.019c1.721,0.218 2.123,0.558 3.065,1.443l51.111,51.11c0.318,0.339 0.566,0.719 0.751,1.126l0.335,0.504l0.378,1.914l-0.384,1.912c-2.542,3.792 -6.593,6.957 -10.411,10.165c-43.029,36.155 -107.286,45.572 -159.303,19.697c-13.554,-6.742 -25.897,-15.671 -36.633,-26.27c-0.292,-0.232 -0.557,-0.494 -0.787,-0.782Zm177.733,-22.445c-1.673,1.564 -3.576,3.103 -5.368,4.55c-30.91,24.96 -74.943,32.906 -112.76,17.18c-11.313,-4.704 -21.735,-11.39 -30.847,-19.572l-17.497,17.497c26.413,23.747 62.281,36.873 98.408,34.767c31.602,-1.843 62.332,-15.13 85.707,-36.778l-17.643,-17.644Zm-26.461,-26.46c-0.976,0.846 -2.007,1.662 -3.007,2.432c-24.638,18.957 -62.085,19.733 -87.789,0.255c-1.094,-0.829 -2.162,-1.691 -3.203,-2.583l-21.552,21.551l0.069,0.06c22.951,20.122 55.428,29.209 85.678,22.83c18.209,-3.84 35.443,-12.829 49.145,-25.204l-19.341,-19.341Zm-159.621,-48.825l-24.862,0c1.479,31.868 14.048,63.442 36.907,87.572c0.129,0.135 0.257,0.27 0.386,0.405l17.537,-17.537c-15.171,-16.25 -25.652,-36.987 -28.931,-59.32c-0.541,-3.688 -0.885,-7.401 -1.037,-11.12Zm1500.15,-82.447c-36.511,0 -75.81,24.001 -75.81,76.7c0,52.815 39.299,76.871 75.81,76.871c36.512,0 75.81,-24.056 75.81,-76.871c0,-52.699 -39.298,-76.7 -75.81,-76.7Zm-1049.9,76.701c0,52.814 39.297,76.87 75.809,76.87c36.514,0 75.811,-24.056 75.811,-76.87c0,-52.7 -39.297,-76.701 -75.811,-76.701c-36.512,0 -75.809,24.001 -75.809,76.701Zm-57.482,-35.201l22.6,-22.599c-13.89,-12.313 -32.804,-18.899 -52.028,-18.899c-10.567,0 -21.227,1.988 -31.206,6.104c-28.273,11.667 -44.487,37.396 -44.487,70.593c0,33.307 16.244,59.105 44.565,70.781c28.109,11.586 61.638,6.246 83.133,-12.837l-22.53,-22.528c-11.692,10.852 -29.079,14.311 -44.648,8.516c-17.499,-6.506 -27.945,-22.928 -27.945,-43.932c0,-20.755 10.372,-37.031 27.745,-43.541c15.61,-5.846 33.051,-2.462 44.801,8.342Zm291.003,109.085l-33.295,0l0,-148.579l19.72,0l73.744,84.017l-0.191,-83.044l33.106,0l0,147.606l-19.528,0l-73.556,-82.418l0,82.418Zm242.751,-147.91l0,27.223l-38.124,0l0,120.384l-31.01,0l0,-120.384l-38.345,0l0,-27.223l107.479,0Zm74.014,-0.001l0,147.608l-31.015,0l0,-147.608l31.015,0Zm130.028,-2.006l64.239,149.614l-33.296,0l-7.561,-19.619l-67.845,0l-7.56,19.619l-33.295,0l64.238,-149.614l21.08,0Zm174.683,72.737l51.929,-71.955l20.723,0l0,148.05l-32.35,0l0.052,-80.279l-36.77,46.53l-6.8,0l-36.744,-46.53l0,80.279l-32.35,0l0,-148.05l20.724,0l51.586,71.955Zm-1264.46,9.042l-30.284,0c1.147,22.858 9.795,45.62 26.402,63.507l21.613,-21.613c-10.14,-11.566 -16.557,-26.423 -17.731,-41.894Zm528.734,-5.746c0,30.655 -21.805,46.694 -43.348,46.694c-21.54,0 -43.347,-16.039 -43.347,-46.694c0,-30.434 21.807,-46.359 43.347,-46.359c21.543,0 43.348,15.925 43.348,46.359Zm930.742,46.694c-21.54,0 -43.349,-16.039 -43.349,-46.695c0,-30.433 21.809,-46.359 43.349,-46.359c21.54,0 43.349,15.926 43.349,46.359c0,30.656 -21.809,46.695 -43.349,46.695Zm-355.927,-18.392l-24.581,-55.835l-23.973,55.835l48.554,0Zm-939.999,-126.48c-0.422,-0.401 -0.848,-0.799 -1.278,-1.189c-38.329,-34.694 -96.652,-46.086 -145.353,-25.477c-36.914,15.621 -66.175,48.285 -77.459,87.024c-3.178,10.915 -4.901,22.211 -5.204,33.566l25.015,0c1.343,-26.307 12.107,-52.181 30.551,-71.422c20.338,-21.217 49.456,-33.789 78.99,-34.018c0.352,-0.001 0.705,-0.002 1.057,-0.001c27.59,0.138 54.653,11.136 75.223,29.975l18.458,-18.458Zm-25.541,25.541c-23.22,-20.96 -56.015,-31.199 -87.184,-25.319c-25.374,4.787 -48.633,19.859 -63.554,41.074c-10.779,15.326 -16.725,33.771 -17.58,52.628l30.291,0c1.976,-26.808 20.269,-51.774 46.23,-61.538c8.272,-3.111 17.053,-4.646 25.98,-4.619c16.874,0.151 33.581,6.19 46.436,17.155l19.381,-19.381Z" />
+      </svg>
+    ),
+    aspectRatio: 1740 / 360,
+  },
+  ContiamoStacked: {
+    svg: (
+      <svg viewBox="0 0 860 560">
+        <path d="M786.686,424.188c-22.908,0 -47.565,15.058 -47.565,48.123c0,33.137 24.657,48.23 47.565,48.23c22.908,0 47.564,-15.093 47.564,-48.23c0,-33.065 -24.656,-48.123 -47.564,-48.123Zm-658.728,48.123c0,33.137 24.655,48.23 47.564,48.23c22.909,0 47.565,-15.093 47.565,-48.23c0,-33.065 -24.656,-48.124 -47.565,-48.124c-22.909,0 -47.564,15.059 -47.564,48.124Zm-36.066,-22.086l14.18,-14.179c-8.715,-7.725 -20.582,-11.857 -32.643,-11.857c-6.63,0 -13.318,1.247 -19.579,3.829c-17.74,7.321 -27.913,23.464 -27.913,44.292c0,20.897 10.192,37.084 27.962,44.409c17.635,7.27 38.672,3.919 52.159,-8.054l-14.136,-14.134c-7.336,6.809 -18.245,8.979 -28.013,5.343c-10.979,-4.082 -17.534,-14.386 -17.534,-27.564c0,-13.022 6.508,-23.234 17.409,-27.318c9.793,-3.668 20.736,-1.545 28.108,5.233Zm182.581,68.443l-20.89,0l0,-93.222l12.373,0l46.269,52.714l-0.12,-52.103l20.771,0l0,92.611l-12.252,0l-46.151,-51.711l0,51.711Zm152.307,-75.722l-23.92,0l0,75.532l-19.456,0l0,-75.532l-24.058,0l0,-17.08l67.434,0l0,17.08Zm46.438,75.531l-19.459,0l0,-92.611l19.459,0l0,92.611Zm121.887,0l-20.891,0l-4.744,-12.309l-42.567,0l-4.743,12.309l-20.89,0l40.304,-93.87l13.226,0l40.305,93.87Zm44.223,-0.49l-20.297,0l0,-92.89l13.003,0l32.366,45.146l32.581,-45.146l13.002,0l0,92.89l-20.297,0l0.033,-50.368l-23.071,29.193l-4.266,0l-23.054,-29.193l0,50.368Zm-436.538,-45.676c0,19.234 -13.681,29.297 -27.198,29.297c-13.515,0 -27.196,-10.063 -27.196,-29.297c0,-19.094 13.681,-29.086 27.196,-29.086c13.517,0 27.198,9.992 27.198,29.086Zm583.965,29.297c-13.514,0 -27.197,-10.063 -27.197,-29.297c0,-19.095 13.683,-29.087 27.197,-29.087c13.515,0 27.199,9.992 27.199,29.087c0,19.234 -13.684,29.297 -27.199,29.297Zm-223.315,-11.54l-15.423,-35.031l-15.041,35.031l30.464,0Zm-215.924,-205.158c-20.365,-20.281 -35.018,-46.456 -40.62,-75.146c-9.268,-47.475 6.874,-99.038 41.839,-132.753c26.739,-25.783 63.27,-40.643 100.587,-40.823c0.308,-0.001 0.615,-0.001 0.923,0c38.168,0.192 76.224,16.349 103.299,42.915c0.504,0.322 0.778,0.725 1.131,1.154c1.579,1.917 1.705,4.547 -0.316,6.709l-0.001,0.001l-0.001,0.001l-0.005,0.005l-51.919,51.919c-0.841,0.788 -1.202,0.944 -1.837,1.167c-0.599,0.23 -1.28,0.337 -2.038,0.286c-1.405,-0.096 -2.527,-0.805 -3.725,-1.969c-0.272,-0.261 -0.272,-0.261 -0.546,-0.519c-22.63,-21.138 -61.591,-22.112 -85.442,0.595c-13.149,12.52 -20.324,30.943 -18.677,49.467c2.403,27.015 24.947,51.187 53.027,55.081c18.455,2.559 38.081,-3.614 51.563,-16.433c0.146,-0.139 0.278,-0.265 0.398,-0.379c0.296,-0.372 0.648,-0.702 1.053,-0.977c0.21,-0.177 0.264,-0.189 0.396,-0.241c0.743,-0.406 1.634,-0.641 2.645,-0.641c0.104,0.007 0.209,0.013 0.313,0.02c1.137,0.143 1.876,0.325 3.065,1.442l51.111,51.111l0.007,0.007c0.786,0.839 0.929,1.187 1.162,1.834c0.145,0.404 0.219,0.817 0.23,1.231c0.135,0.728 0.124,1.322 -0.319,2.386c-0.437,1.048 -1.266,1.804 -2.043,2.567c-37.923,37.239 -96.303,52.114 -147.961,35.345c-21.004,-6.819 -40.191,-18.705 -56.037,-34.121c-0.492,-0.342 -0.924,-0.759 -1.262,-1.241Zm177.868,-22.221c-22.04,20.148 -51.973,31.634 -81.767,30.027c-24.639,-1.328 -48.548,-11.384 -67.202,-27.807l-17.277,17.277c26.622,23.752 62.523,37.388 99.475,34.879c31.167,-2.117 61.406,-15.38 84.447,-36.7l-17.676,-17.676Zm-120.246,-26.503l-21.65,21.651c22.941,19.887 54.908,29.551 85.974,22.905c18.087,-3.87 35.121,-12.696 48.842,-25.134l-19.356,-19.356c-18.142,15.52 -44.34,21.239 -67.603,14.401c-9.679,-2.846 -18.604,-7.87 -26.207,-14.467Zm-65.713,-48.749l-24.824,0c0.726,19.334 5.475,38.563 14.441,56.037c5.996,11.687 13.758,22.383 22.878,31.788l17.291,-17.291c-6.093,-6.421 -11.447,-13.604 -15.86,-21.445c-8.414,-14.951 -13.13,-31.899 -13.926,-49.089Zm40.696,0l-30.727,0c0.933,17.868 6.564,35.522 16.634,50.637c3.044,4.569 6.475,8.862 10.237,12.844l21.778,-21.778c-10.088,-11.459 -16.717,-26.06 -17.922,-41.703Zm163.554,-103.93c-33.3,-30.81 -81.464,-44.224 -125.718,-33.556c-32.117,7.742 -61.077,27.739 -79.898,55.033c-14.624,21.207 -22.603,46.668 -23.476,72.453l24.823,0c0.498,-11.745 2.84,-23.444 7.097,-34.538c15.75,-41.051 57.424,-70.552 102.429,-70.901c0.352,-0.002 0.705,-0.003 1.058,-0.002c27.59,0.139 54.652,11.136 75.222,29.975l18.463,-18.464Zm-25.547,25.548c-23.219,-20.96 -56.015,-31.199 -87.183,-25.32c-36.073,6.805 -67.442,34.365 -77.93,71.218c-2.088,7.336 -3.298,14.89 -3.65,22.484l30.738,0c1.973,-26.476 20.098,-51.361 45.41,-61.225c8.51,-3.316 17.656,-4.954 26.798,-4.932c16.95,0.152 33.603,6.171 46.441,17.151l19.376,-19.376Z" />
+      </svg>
+    ),
+    aspectRatio: 860 / 560,
+  },
+  Pantheon: {
+    svg: (
+      <svg viewBox="0 0 1760 360">
+        <path d="M182.174,310.103c-1.121,0.083 -1.998,1.032 -1.991,2.156l0.279,11.056c0.018,1.029 0.869,1.865 1.897,1.865c0.054,0 0.107,-0.002 0.16,-0.007l17.214,-1.904c0.932,-0.129 1.633,-0.933 1.633,-1.874c0,-0.131 -0.013,-0.261 -0.04,-0.389l-2.19,-10.844c-0.238,-1.089 -1.284,-1.824 -2.389,-1.679l-14.573,1.62Zm-47.004,9.788c0.187,0.06 0.383,0.09 0.579,0.09c0.861,0 1.619,-0.585 1.837,-1.417l2.734,-10.724c0.26,-1.083 -0.367,-2.196 -1.427,-2.535l-13.85,-4.845c-1.042,-0.387 -2.221,0.096 -2.694,1.102l-4.552,10.087c-0.113,0.247 -0.172,0.516 -0.172,0.789c0,0.785 0.487,1.492 1.22,1.772l16.325,5.681Zm75.751,-14.6c-1.06,0.339 -1.687,1.453 -1.426,2.535l2.734,10.738c0.217,0.832 0.975,1.417 1.836,1.417c0.197,0 0.392,-0.03 0.579,-0.09l16.332,-5.707c0.731,-0.279 1.217,-0.985 1.217,-1.767c0,-0.275 -0.06,-0.546 -0.175,-0.795l-4.553,-10.087c-0.472,-1.006 -1.652,-1.488 -2.694,-1.101l-13.85,4.857Zm-111.394,-18.999c-0.928,-0.614 -2.189,-0.407 -2.873,0.471l-6.676,8.813c-0.254,0.332 -0.392,0.739 -0.392,1.157c0,0.632 0.314,1.223 0.837,1.577l14.599,9.231c0.292,0.173 0.626,0.265 0.965,0.265c0.714,0 1.369,-0.404 1.69,-1.041l5.123,-9.848c0.495,-0.998 0.134,-2.224 -0.823,-2.794l-12.45,-7.831Zm94.439,-1.692c-0.234,-1.092 -1.281,-1.831 -2.389,-1.686l-10.054,1.142c-1.11,0.1 -1.969,1.048 -1.957,2.163l0.259,10.445c0.017,1.025 0.865,1.859 1.891,1.859c0.053,0 0.106,-0.003 0.159,-0.007l12.562,-1.394c0.931,-0.125 1.633,-0.928 1.633,-1.867c0,-0.131 -0.013,-0.261 -0.04,-0.389l-2.064,-10.266Zm-26.033,13.889c0.053,0.005 0.106,0.007 0.159,0.007c1.025,0 1.873,-0.833 1.891,-1.858l0.266,-10.439c0.011,-1.115 -0.847,-2.063 -1.958,-2.163l-10.054,-1.122c-1.108,-0.145 -2.154,0.594 -2.389,1.686l-2.07,10.239c-0.027,0.128 -0.041,0.258 -0.041,0.389c0,0.939 0.703,1.742 1.633,1.868l12.563,1.393Zm-32.159,-21.832c-1.044,-0.388 -2.226,0.094 -2.701,1.101l-4.294,9.53c-0.115,0.248 -0.174,0.518 -0.174,0.792c0,0.782 0.486,1.486 1.216,1.762l11.945,4.175c0.188,0.06 0.383,0.09 0.58,0.09c0.86,0 1.618,-0.585 1.836,-1.418l2.581,-10.126c0.262,-1.083 -0.361,-2.197 -1.42,-2.542l-9.569,-3.364Zm132.762,11.938c0.368,0.442 0.914,0.698 1.49,0.698c0.479,0 0.942,-0.178 1.297,-0.499c0.356,-0.321 3.843,-3.451 6.311,-5.913c2.469,-2.462 5.913,-6.317 5.913,-6.317c0.321,-0.356 0.499,-0.819 0.499,-1.298c0,-0.575 -0.256,-1.122 -0.698,-1.49l-8.468,-7.107c-0.856,-0.704 -2.132,-0.613 -2.88,0.206c-0.748,0.819 -2.86,3.185 -5.023,5.349c-2.164,2.163 -5.342,5.016 -5.342,5.016c-0.82,0.751 -0.911,2.028 -0.206,2.887l7.107,8.468Zm-25.29,-11.56c0.526,-0.351 0.842,-0.942 0.842,-1.574c0,-0.42 -0.14,-0.829 -0.397,-1.16l-6.305,-8.322c-0.682,-0.881 -1.943,-1.09 -2.873,-0.478l-8.561,5.409c-0.955,0.574 -1.313,1.802 -0.816,2.8l4.765,9.291c0.323,0.636 0.978,1.038 1.692,1.038c0.338,0 0.671,-0.09 0.962,-0.262l10.691,-6.742Zm-74.703,-2.933c0.053,0.004 0.106,0.006 0.16,0.006c1.025,0 1.873,-0.833 1.891,-1.858l0.199,-7.963c0.015,-1.116 -0.84,-2.067 -1.951,-2.17l-6.271,-0.717c-1.107,-0.142 -2.152,0.595 -2.389,1.686l-1.593,7.844c-0.027,0.127 -0.041,0.257 -0.041,0.387c0,0.938 0.703,1.74 1.634,1.862l8.361,0.923Zm41.416,-6.497c0.731,-0.279 1.217,-0.985 1.217,-1.767c0,-0.275 -0.06,-0.546 -0.175,-0.795l-3.318,-7.299c-0.475,-1.008 -1.657,-1.49 -2.701,-1.102l-5.972,2.084c-1.058,0.35 -1.679,1.464 -1.421,2.548l1.991,7.771c0.215,0.833 0.972,1.42 1.832,1.42c0.198,0 0.395,-0.032 0.584,-0.093l7.963,-2.767Zm-114.307,0.743c0.356,0.321 0.819,0.499 1.298,0.499c0.575,0 1.121,-0.256 1.489,-0.698l6.709,-7.963c0.707,-0.858 0.619,-2.135 -0.199,-2.887c-0.818,-0.752 -2.044,-1.845 -3.676,-3.477c-1.633,-1.633 -3.484,-3.677 -3.484,-3.677c-0.75,-0.817 -2.025,-0.905 -2.88,-0.199l-7.99,6.676c-0.44,0.368 -0.694,0.913 -0.694,1.486c0,0.481 0.179,0.945 0.501,1.301c0.323,0.357 2.409,2.655 4.334,4.599c1.924,1.944 4.592,4.34 4.592,4.34Zm163.142,-4.34c1.925,-1.924 4.334,-4.599 4.334,-4.599c0.323,-0.356 0.502,-0.82 0.502,-1.301c0,-0.573 -0.255,-1.118 -0.695,-1.486l-8.003,-6.709c-0.855,-0.706 -2.13,-0.618 -2.88,0.199c-0.75,0.817 -1.838,2.037 -3.484,3.676c-1.646,1.64 -3.676,3.478 -3.676,3.478c-0.818,0.752 -0.906,2.029 -0.199,2.886l6.715,7.964c0.366,0.443 0.911,0.7 1.485,0.7c0.479,0 0.942,-0.179 1.296,-0.501c0.354,-0.323 2.648,-2.376 4.605,-4.307Zm31.542,1.049c0.331,0.254 0.738,0.392 1.156,0.392c0.633,0 1.226,-0.317 1.578,-0.844l9.231,-14.599c0.174,-0.292 0.266,-0.626 0.266,-0.965c0,-0.716 -0.408,-1.372 -1.049,-1.69l-9.842,-5.05c-0.998,-0.493 -2.223,-0.136 -2.8,0.817l-7.824,12.402c-0.617,0.926 -0.409,2.187 0.471,2.867l8.813,6.67Zm-229.398,-21.959c-0.577,-0.953 -1.802,-1.31 -2.8,-0.817l-9.841,5.05c-0.641,0.32 -1.047,0.977 -1.047,1.692c0,0.339 0.091,0.671 0.263,0.963l9.231,14.599c0.352,0.527 0.945,0.844 1.579,0.844c0.418,0 0.824,-0.138 1.155,-0.392l8.82,-6.683c0.875,-0.684 1.079,-1.942 0.464,-2.867l-7.824,-12.389Zm67.636,4.884c-0.934,-0.61 -2.194,-0.401 -2.88,0.478l-4.851,6.397c-0.255,0.331 -0.393,0.738 -0.393,1.156c0,0.633 0.317,1.226 0.844,1.578l7.101,4.479c0.292,0.175 0.627,0.268 0.967,0.268c0.714,0 1.369,-0.405 1.687,-1.044l3.65,-7.147c0.491,-1.001 0.135,-2.226 -0.816,-2.807l-5.309,-3.358Zm99.946,9.609c0.527,-0.352 0.844,-0.945 0.844,-1.578c0,-0.418 -0.138,-0.825 -0.392,-1.156l-4.851,-6.397c-0.685,-0.882 -1.948,-1.092 -2.88,-0.478l-5.309,3.358c-0.951,0.581 -1.308,1.806 -0.817,2.807l3.67,7.147c0.319,0.639 0.973,1.044 1.687,1.044c0.341,0 0.675,-0.093 0.968,-0.268l7.08,-4.479Zm1202.16,-0.66c-14.534,0 -27.634,-3.321 -39.3,-9.964c-11.667,-6.642 -20.8,-15.794 -27.4,-27.454c-6.6,-11.66 -9.9,-24.768 -9.9,-39.326c0,-14.557 3.3,-27.666 9.9,-39.326c6.6,-11.66 15.733,-20.811 27.4,-27.454c11.666,-6.642 24.766,-9.964 39.3,-9.964c14.533,0 27.6,3.322 39.2,9.964c11.6,6.643 20.733,15.794 27.4,27.454c6.666,11.66 10,24.769 10,39.326c0,14.558 -3.334,27.666 -10,39.326c-6.667,11.66 -15.8,20.812 -27.4,27.454c-11.6,6.643 -24.667,9.964 -39.2,9.964Zm-688.4,-2.544l-32,0l0,-148.4l26.8,0l69.6,90.1l0,-90.1l32,0l0,148.4l-26.6,0l-69.8,-90.1l0,90.1Zm321.6,0l-32.4,0l0,-148.4l32.4,0l0,58.512l63.6,0l0,-58.512l32.4,0l0,148.4l-32.4,0l0,-60.844l-63.6,0l0,60.844Zm517.8,0l-32,0l0,-148.4l26.8,0l69.6,90.1l0,-90.1l32,0l0,148.4l-26.6,0l-69.8,-90.1l0,90.1Zm-630.4,0l-32.4,0l0,-120.416l-44.8,0l0,-27.984l122,0l0,27.984l-44.8,0l0,120.416Zm368.2,0l-108.4,0l0,-148.4l105.8,0l0,27.56l-73.6,0l0,32.224l65,0l0,26.712l-65,0l0,34.344l76.2,0l0,27.56Zm-641.8,0l-34,0l-12.4,-31.8l-65,0l-12.4,31.8l-33.2,0l62.4,-148.4l32,0l62.6,148.4Zm-261.6,0l-32.4,0l0,-148.4l60.6,0c12.4,0 23.166,2.191 32.3,6.572c9.133,4.382 16.166,10.6 21.1,18.656c4.933,8.056 7.4,17.596 7.4,28.62c0,10.883 -2.467,20.388 -7.4,28.514c-4.934,8.127 -11.967,14.346 -21.1,18.656c-9.134,4.311 -19.9,6.466 -32.3,6.466l-28.2,0l0,40.916Zm-230.029,-8.96c-0.237,-1.092 -1.28,-1.831 -2.389,-1.692l-3.139,0.351c-1.11,0.11 -1.963,1.061 -1.951,2.177l0.146,5.76c0.014,1.017 0.854,1.847 1.872,1.847c0.059,0 0.119,-0.003 0.179,-0.009l4.831,-0.537c0.931,-0.125 1.633,-0.929 1.633,-1.868c0,-0.13 -0.013,-0.261 -0.04,-0.388l-1.142,-5.641Zm-33.824,-5.236c-1.048,-0.382 -2.228,0.105 -2.701,1.115l-2.369,5.249c-0.114,0.247 -0.174,0.516 -0.174,0.789c0,0.782 0.489,1.486 1.222,1.759l4.593,1.613c0.188,0.061 0.384,0.092 0.582,0.092c0.859,0 1.615,-0.587 1.827,-1.419l1.42,-5.581c0.259,-1.086 -0.362,-2.202 -1.42,-2.555l-2.98,-1.062Zm49.386,8.912c0.733,-0.273 1.222,-0.977 1.222,-1.759c0,-0.273 -0.059,-0.542 -0.174,-0.789l-2.362,-5.249c-0.476,-1.011 -1.658,-1.498 -2.708,-1.115l-2.979,1.049c-1.059,0.352 -1.679,1.469 -1.42,2.554l1.42,5.581c0.211,0.834 0.967,1.422 1.827,1.422c0.2,0 0.398,-0.032 0.588,-0.094l4.586,-1.6Zm-164.815,-31.594c-0.991,0.243 -1.693,1.139 -1.693,2.159c0,0.231 0.036,0.461 0.107,0.681l5.482,15.694c0.33,0.856 1.157,1.423 2.074,1.423c0.319,0 0.635,-0.069 0.925,-0.202l9.47,-4.306c1.18,-0.554 1.744,-1.938 1.287,-3.159l-4.612,-13.213c-0.404,-1.24 -1.711,-1.971 -2.979,-1.665l-10.061,2.588Zm91.652,6.875c-0.777,-0.802 -2.054,-0.878 -2.92,-0.173l-4.407,3.697c-0.433,0.356 -0.685,0.889 -0.685,1.45c0,0.486 0.189,0.954 0.526,1.304l3.431,3.437c0.35,0.336 0.816,0.523 1.3,0.523c0.563,0 1.098,-0.253 1.454,-0.689l3.696,-4.4c0.707,-0.864 0.635,-2.141 -0.166,-2.919l-2.229,-2.23Zm100.404,6.278c0.336,-0.351 0.523,-0.818 0.523,-1.303c0,-0.561 -0.25,-1.094 -0.682,-1.451l-4.406,-3.69c-0.865,-0.707 -2.142,-0.634 -2.92,0.166l-2.23,2.23c-0.8,0.778 -0.873,2.055 -0.166,2.919l3.696,4.407c0.358,0.432 0.891,0.682 1.452,0.682c0.485,0 0.952,-0.188 1.302,-0.523l3.431,-3.437Zm20.386,-10.034c-1.179,-0.597 -2.637,-0.177 -3.318,0.955l-2.966,4.739c-0.715,1.089 -0.471,2.563 0.557,3.364l5.86,4.44c0.388,0.299 0.865,0.462 1.355,0.462c0.74,0 1.433,-0.371 1.844,-0.987l4.114,-6.523c0.206,-0.343 0.315,-0.735 0.315,-1.135c0,-0.833 -0.472,-1.598 -1.217,-1.971l-6.544,-3.344Zm1181.44,7.047c8.266,0 15.733,-2.014 22.4,-6.042c6.666,-4.028 11.9,-9.646 15.7,-16.854c3.8,-7.208 5.7,-15.405 5.7,-24.592c0,-9.186 -1.9,-17.384 -5.7,-24.592c-3.8,-7.208 -9.034,-12.826 -15.7,-16.854c-6.667,-4.028 -14.134,-6.042 -22.4,-6.042c-8.267,0 -15.734,2.014 -22.4,6.042c-6.667,4.028 -11.9,9.646 -15.7,16.854c-3.8,7.208 -5.7,15.406 -5.7,24.592c0,9.187 1.9,17.384 5.7,24.592c3.8,7.208 9.033,12.826 15.7,16.854c6.666,4.028 14.133,6.042 22.4,6.042Zm-1153.01,-19.171c-1.269,-0.306 -2.576,0.424 -2.98,1.665l-3.112,8.919c-0.46,1.223 0.108,2.61 1.294,3.159l8.905,4.021c0.29,0.134 0.606,0.203 0.925,0.203c0.918,0 1.745,-0.571 2.068,-1.43l3.949,-11.282c0.07,-0.218 0.105,-0.446 0.105,-0.675c0,-1.016 -0.698,-1.908 -1.685,-2.152l-9.469,-2.428Zm-102.451,-74.236c21.135,0.205 40.813,16.446 44.801,37.235c4.341,22.63 -10.475,47.664 -34.458,53.363c-19.985,4.748 -42.704,-5.455 -52.025,-24.772c-8.388,-17.384 -4.216,-40.042 9.911,-53.333c8.494,-7.991 19.906,-12.527 31.771,-12.493Zm-54.56,84.555c1.031,-0.801 1.28,-2.276 0.571,-3.371l-1.328,-2.091c-0.685,-1.128 -2.14,-1.544 -3.318,-0.949l-4.512,2.31c-0.742,0.371 -1.212,1.133 -1.212,1.963c0,0.4 0.109,0.793 0.316,1.136l2.23,3.537c0.41,0.616 1.104,0.987 1.844,0.987c0.487,0 0.961,-0.16 1.348,-0.456l4.061,-3.066Zm-23.671,-5.614c1.193,-0.542 1.775,-1.928 1.327,-3.159l-1.858,-5.309c-0.41,-1.24 -1.716,-1.968 -2.986,-1.666l-7.114,1.812c-0.984,0.239 -1.683,1.128 -1.683,2.141c0,0.233 0.037,0.464 0.11,0.686l2.542,7.3c0.325,0.855 1.15,1.425 2.065,1.425c0.321,0 0.637,-0.07 0.927,-0.204l6.67,-3.026Zm77.702,-68.941c-17.466,0.169 -33.601,14.513 -35.478,32.014c-2.099,19.57 14.189,39.691 35.478,39.825c16.047,0.101 31.402,-11.744 35.183,-27.595c3.785,-15.871 -4.662,-33.829 -19.781,-40.89c-4.798,-2.24 -10.129,-3.369 -15.402,-3.354Zm64.198,54.368c-1.27,-0.307 -2.58,0.419 -2.993,1.659l-0.823,2.342c-0.445,1.231 0.137,2.615 1.328,3.159l4.645,2.09c0.29,0.134 0.605,0.203 0.925,0.203c0.918,0 1.744,-0.571 2.068,-1.43l1.38,-3.949c0.073,-0.221 0.11,-0.453 0.11,-0.686c0,-1.009 -0.696,-1.895 -1.676,-2.134l-4.964,-1.254Zm68.531,-13.054c-1.305,-0.018 -2.417,0.989 -2.528,2.29l-1.546,13.896c-0.172,1.302 0.703,2.531 1.991,2.794l10.173,2.057c0.149,0.031 0.301,0.047 0.454,0.047c1.1,0 2.043,-0.817 2.2,-1.905l1.832,-16.518c0.005,-0.059 0.007,-0.119 0.007,-0.179c0,-1.215 -0.996,-2.216 -2.21,-2.223l-10.373,-0.259Zm-249.551,13.883c0.147,1.091 1.089,1.914 2.19,1.914c0.156,0 0.312,-0.016 0.464,-0.049l9.576,-1.938c1.286,-0.265 2.16,-1.492 1.991,-2.794l-1.095,-9.383c-0.12,-1.298 -1.231,-2.299 -2.535,-2.283l-9.742,0.246c-1.201,0.02 -2.178,1.014 -2.178,2.216c0,0.062 0.003,0.124 0.008,0.186l1.321,11.885Zm201.539,-15.084c-1.307,-0.016 -2.421,0.988 -2.542,2.29l-0.623,5.581c-0.166,1.301 0.707,2.525 1.99,2.793l7.174,1.447c0.155,0.034 0.313,0.051 0.471,0.051c1.098,0 2.037,-0.821 2.183,-1.909l0.85,-7.665c0.005,-0.062 0.008,-0.124 0.008,-0.186c0,-1.199 -0.973,-2.192 -2.171,-2.216l-7.34,-0.186Zm361.673,12.13l-22.2,-56.816l-22.2,56.816l44.4,0Zm-178.6,-11.024c9.733,0 17.133,-2.226 22.2,-6.678c5.066,-4.452 7.6,-10.776 7.6,-18.974c0,-8.338 -2.534,-14.734 -7.6,-19.186c-5.067,-4.452 -12.467,-6.678 -22.2,-6.678l-26.4,0l0,51.516l26.4,0Zm-352.069,-9.348c1.308,0.016 2.422,-0.987 2.542,-2.289l0.624,-5.581c0.165,-1.301 -0.707,-2.526 -1.991,-2.794l-7.174,-1.447c-0.154,-0.034 -0.312,-0.051 -0.471,-0.051c-1.098,0 -2.037,0.821 -2.183,1.909l-0.81,7.665c-0.005,0.062 -0.008,0.124 -0.008,0.186c0,1.202 0.977,2.195 2.178,2.216l7.293,0.186Zm200.69,-0.796c1.201,-0.021 2.178,-1.015 2.178,-2.216c0,-0.062 -0.003,-0.125 -0.008,-0.186l-1.327,-11.886c-0.147,-1.091 -1.089,-1.914 -2.19,-1.914c-0.156,0 -0.312,0.017 -0.465,0.049l-9.576,1.938c-1.287,0.263 -2.162,1.492 -1.99,2.794l1.042,9.384c0.117,1.297 1.225,2.298 2.528,2.282l9.808,-0.245Zm-256.765,-21.442c-0.149,-0.031 -0.301,-0.046 -0.454,-0.046c-1.1,0 -2.043,0.816 -2.2,1.905l-1.832,16.517c-0.005,0.06 -0.007,0.12 -0.007,0.179c0,1.205 0.979,2.202 2.184,2.223l10.379,0.259c1.301,0.015 2.408,-0.987 2.521,-2.283l1.547,-13.903c0.172,-1.302 -0.704,-2.531 -1.991,-2.793l-10.147,-2.058Zm76.568,7.984c1.271,0.304 2.581,-0.42 2.999,-1.659l0.817,-2.343c0.444,-1.231 -0.137,-2.615 -1.328,-3.159l-4.572,-2.063c-0.29,-0.134 -0.606,-0.203 -0.925,-0.203c-0.918,0 -1.744,0.571 -2.068,1.43l-1.38,3.949c-0.073,0.221 -0.11,0.453 -0.11,0.686c0,1.009 0.695,1.895 1.676,2.134l4.891,1.228Zm141.92,-14.573c-1.193,0.541 -1.776,1.927 -1.327,3.159l1.858,5.309c0.41,1.239 1.716,1.968 2.986,1.665l7.114,-1.811c0.984,-0.24 1.682,-1.128 1.682,-2.141c0,-0.233 -0.037,-0.465 -0.109,-0.686l-2.542,-7.3c-0.325,-0.856 -1.15,-1.425 -2.066,-1.425c-0.32,0 -0.636,0.069 -0.927,0.204l-6.669,3.026Zm-180.675,4.685c1.269,0.305 2.575,-0.425 2.979,-1.666l3.113,-8.919c0.459,-1.222 -0.109,-2.61 -1.294,-3.159l-8.906,-4.021c-0.29,-0.134 -0.606,-0.203 -0.925,-0.203c-0.918,0 -1.744,0.572 -2.068,1.431l-3.982,11.281c-0.069,0.218 -0.104,0.445 -0.104,0.674c0,1.018 0.701,1.912 1.69,2.153l9.497,2.429Zm227.838,-8.09c0.402,1.24 1.706,1.971 2.973,1.666l10.06,-2.561c0.989,-0.245 1.689,-1.14 1.689,-2.158c0,-0.227 -0.034,-0.453 -0.103,-0.669l-5.481,-15.695c-0.33,-0.855 -1.157,-1.423 -2.074,-1.423c-0.32,0 -0.635,0.069 -0.926,0.202l-9.469,4.267c-1.181,0.554 -1.745,1.938 -1.288,3.159l4.619,13.212Zm-199.409,-4.014c1.171,0.576 2.606,0.159 3.285,-0.956l2.999,-4.751c0.716,-1.09 0.471,-2.564 -0.557,-3.365l-5.86,-4.44c-0.388,-0.299 -0.864,-0.462 -1.355,-0.462c-0.74,0 -1.433,0.371 -1.843,0.987l-4.115,6.523c-0.206,0.343 -0.315,0.735 -0.315,1.135c0,0.833 0.473,1.598 1.218,1.971l6.543,3.358Zm120.791,-3.975c0.777,0.802 2.054,0.877 2.92,0.172l4.406,-3.696c0.432,-0.358 0.682,-0.891 0.682,-1.451c0,-0.486 -0.187,-0.953 -0.523,-1.303l-3.431,-3.438c-0.349,-0.335 -0.815,-0.523 -1.3,-0.523c-0.563,0 -1.097,0.253 -1.454,0.689l-3.689,4.393c-0.707,0.865 -0.635,2.141 0.166,2.92l2.223,2.237Zm-94.22,-9.55c-0.356,-0.436 -0.891,-0.689 -1.454,-0.689c-0.484,0 -0.95,0.188 -1.3,0.523l-3.431,3.438c-0.337,0.35 -0.526,0.817 -0.526,1.303c0,0.562 0.252,1.094 0.685,1.451l4.407,3.696c0.866,0.705 2.143,0.63 2.92,-0.172l2.229,-2.23c0.801,-0.779 0.873,-2.056 0.166,-2.92l-3.696,-4.4Zm83.204,0.312c0.935,0.609 2.195,0.397 2.88,-0.484l3.47,-4.579c0.257,-0.331 0.396,-0.737 0.396,-1.155c0,-0.632 -0.319,-1.224 -0.847,-1.573l-4.114,-2.594c-0.294,-0.178 -0.63,-0.272 -0.973,-0.272c-0.712,0 -1.365,0.404 -1.682,1.041l-2.654,5.123c-0.496,0.999 -0.142,2.226 0.81,2.808l2.714,1.685Zm-55.146,-15.754c-0.212,-0.832 -0.968,-1.42 -1.827,-1.42c-0.198,0 -0.394,0.032 -0.582,0.093l-4.593,1.612c-0.732,0.276 -1.219,0.981 -1.219,1.764c0,0.271 0.058,0.538 0.171,0.785l2.369,5.249c0.473,1.01 1.653,1.497 2.701,1.115l2.98,-1.049c1.058,-0.353 1.679,-1.47 1.42,-2.555l-1.42,-5.594Zm133.671,6.278c0.575,0.955 1.802,1.313 2.8,0.816l9.842,-5.043c0.64,-0.32 1.046,-0.977 1.046,-1.693c0,-0.338 -0.091,-0.671 -0.263,-0.962l-9.231,-14.599c-0.35,-0.526 -0.942,-0.843 -1.574,-0.843c-0.42,0 -0.828,0.14 -1.16,0.398l-8.813,6.676c-0.88,0.68 -1.088,1.941 -0.471,2.867l7.824,12.383Zm-229.398,-21.953c-0.331,-0.257 -0.74,-0.397 -1.16,-0.397c-0.632,0 -1.223,0.316 -1.574,0.842l-9.231,14.6c-0.17,0.291 -0.26,0.622 -0.26,0.959c0,0.716 0.405,1.373 1.044,1.695l9.841,5.043c0.998,0.497 2.226,0.14 2.8,-0.816l7.824,-12.409c0.615,-0.925 0.411,-2.184 -0.464,-2.867l-8.82,-6.65Zm121.289,11.189c-0.058,-0.005 -0.115,-0.008 -0.173,-0.008c-1.021,0 -1.864,0.832 -1.878,1.853l-0.146,5.753c-0.012,1.116 0.841,2.067 1.951,2.177l3.139,0.352c1.109,0.141 2.155,-0.599 2.389,-1.693l1.142,-5.64c0.027,-0.128 0.04,-0.258 0.04,-0.388c0,-0.938 -0.703,-1.74 -1.633,-1.862l-4.831,-0.544Zm-16.551,0.544c-0.93,0.122 -1.633,0.924 -1.633,1.862c0,0.13 0.014,0.26 0.041,0.388l1.141,5.64c0.235,1.094 1.28,1.834 2.389,1.693l3.139,-0.352c1.11,-0.11 1.963,-1.061 1.951,-2.177l-0.146,-5.753c-0.014,-1.021 -0.857,-1.853 -1.878,-1.853c-0.057,0 -0.115,0.003 -0.173,0.008l-4.831,0.544Zm-42.922,-4.267c-0.526,0.35 -0.843,0.942 -0.843,1.574c0,0.417 0.137,0.823 0.392,1.153l4.851,6.404c0.686,0.88 1.949,1.086 2.88,0.472l5.309,-3.358c0.951,-0.582 1.307,-1.807 0.816,-2.807l-3.67,-7.148c-0.317,-0.637 -0.97,-1.041 -1.681,-1.041c-0.343,0 -0.68,0.094 -0.973,0.272l-7.081,4.479Zm99.947,9.603c0.93,0.614 2.193,0.408 2.88,-0.472l4.851,-6.404c0.254,-0.33 0.392,-0.736 0.392,-1.153c0,-0.632 -0.317,-1.224 -0.844,-1.574l-7.1,-4.473c-0.294,-0.177 -0.63,-0.271 -0.973,-0.271c-0.712,0 -1.365,0.404 -1.682,1.041l-3.669,7.147c-0.491,1.001 -0.135,2.226 0.816,2.807l5.329,3.352Zm-130.195,-16.027c-1.924,1.925 -4.333,4.606 -4.333,4.606c-0.322,0.355 -0.501,0.818 -0.501,1.298c0,0.572 0.254,1.116 0.694,1.482l7.976,6.716c0.856,0.703 2.129,0.615 2.88,-0.199c0.752,-0.814 1.845,-2.044 3.484,-3.683c1.639,-1.639 3.677,-3.477 3.677,-3.477c0.818,-0.753 0.906,-2.03 0.199,-2.887l-6.709,-7.997c-0.367,-0.441 -0.912,-0.697 -1.486,-0.697c-0.481,0 -0.946,0.18 -1.302,0.505c-0.355,0.324 -2.647,2.402 -4.579,4.333Zm163.103,-4.333c-0.354,-0.324 -0.818,-0.504 -1.298,-0.504c-0.573,0 -1.117,0.255 -1.482,0.696l-6.716,7.997c-0.707,0.857 -0.619,2.134 0.199,2.887c0.818,0.752 2.044,1.844 3.676,3.477c1.633,1.632 3.498,3.676 3.498,3.676c0.751,0.814 2.023,0.902 2.88,0.199l8.003,-6.715c0.439,-0.367 0.694,-0.911 0.694,-1.483c0,-0.48 -0.179,-0.943 -0.502,-1.298c-0.322,-0.355 -2.409,-2.654 -4.333,-4.605c-1.925,-1.951 -4.619,-4.327 -4.619,-4.327Zm-103.948,-0.684c-0.218,-0.832 -0.976,-1.417 -1.836,-1.417c-0.197,0 -0.392,0.03 -0.58,0.09l-7.93,2.774c-0.731,0.274 -1.219,0.977 -1.219,1.759c0,0.275 0.061,0.547 0.177,0.796l3.318,7.3c0.473,1.01 1.655,1.495 2.701,1.108l5.973,-2.084c1.06,-0.35 1.682,-1.469 1.42,-2.555l-2.024,-7.771Zm9.994,5.94c0.231,1.089 1.27,1.829 2.376,1.692l6.257,-0.704c1.113,-0.1 1.97,-1.053 1.952,-2.17l-0.2,-7.963c-0.018,-1.025 -0.866,-1.858 -1.891,-1.858c-0.053,0 -0.106,0.002 -0.159,0.007l-8.342,0.929c-0.93,0.121 -1.634,0.923 -1.634,1.862c0,0.132 0.014,0.264 0.042,0.394l1.599,7.811Zm21.076,-11.023c-0.053,-0.004 -0.106,-0.007 -0.159,-0.007c-1.025,0 -1.873,0.834 -1.891,1.859l-0.199,7.963c-0.019,1.116 0.838,2.07 1.951,2.17l6.257,0.703c1.108,0.146 2.155,-0.593 2.389,-1.685l1.586,-7.837c0.028,-0.13 0.042,-0.262 0.042,-0.395c0,-0.938 -0.703,-1.74 -1.634,-1.862l-8.342,-0.909Zm53.242,8.607c0.929,0.611 2.187,0.405 2.873,-0.471l6.305,-8.328c0.257,-0.332 0.397,-0.741 0.397,-1.161c0,-0.631 -0.316,-1.223 -0.842,-1.573l-10.691,-6.743c-0.292,-0.175 -0.627,-0.267 -0.968,-0.267c-0.713,0 -1.368,0.405 -1.686,1.044l-4.765,9.29c-0.497,0.998 -0.139,2.226 0.816,2.801l8.561,5.408Zm-127.945,-11.533c-0.525,0.35 -0.842,0.942 -0.842,1.573c0,0.42 0.14,0.829 0.398,1.161l6.304,8.328c0.686,0.876 1.944,1.082 2.873,0.471l8.561,-5.402c0.955,-0.575 1.313,-1.802 0.816,-2.8l-4.764,-9.291c-0.319,-0.638 -0.974,-1.044 -1.687,-1.044c-0.341,0 -0.675,0.093 -0.968,0.268l-10.691,6.736Zm160.409,4.817c2.163,2.164 5.023,5.342 5.023,5.342c0.748,0.82 2.024,0.911 2.88,0.206l8.468,-7.107c0.442,-0.368 0.698,-0.914 0.698,-1.489c0,-0.48 -0.178,-0.942 -0.499,-1.298c-0.321,-0.356 -3.451,-3.856 -5.913,-6.311c-2.462,-2.456 -6.311,-5.913 -6.311,-5.913c-0.356,-0.323 -0.82,-0.502 -1.301,-0.502c-0.573,0 -1.118,0.255 -1.486,0.694l-7.087,8.475c-0.703,0.856 -0.615,2.128 0.199,2.88c0.814,0.751 3.159,2.853 5.329,5.023Zm-133.904,-5.541c0.473,1.01 1.657,1.493 2.701,1.102l9.543,-3.318c1.058,-0.345 1.681,-1.46 1.42,-2.542l-2.582,-10.127c-0.217,-0.832 -0.975,-1.417 -1.836,-1.417c-0.196,0 -0.392,0.03 -0.579,0.09l-11.945,4.174c-0.732,0.277 -1.219,0.981 -1.219,1.764c0,0.275 0.06,0.548 0.177,0.798l4.32,9.476Zm80.967,1.102c1.045,0.391 2.228,-0.092 2.701,-1.102l4.294,-9.516c0.117,-0.25 0.177,-0.522 0.177,-0.798c0,-0.783 -0.487,-1.487 -1.219,-1.764l-11.945,-4.174c-0.187,-0.06 -0.383,-0.09 -0.579,-0.09c-0.861,0 -1.619,0.585 -1.837,1.418l-2.581,10.146c-0.261,1.082 0.361,2.197 1.42,2.542l9.569,3.338Zm-58.192,-7.964c0.238,1.09 1.283,1.825 2.389,1.679l10.054,-1.114c1.109,-0.103 1.966,-1.05 1.958,-2.164l-0.266,-10.438c-0.014,-1.028 -0.863,-1.865 -1.891,-1.865c-0.053,0 -0.106,0.002 -0.159,0.006l-12.563,1.394c-0.931,0.125 -1.634,0.928 -1.634,1.868c0,0.133 0.014,0.265 0.042,0.395l2.07,10.239Zm-43.891,-9.502c0.957,-0.571 1.318,-1.796 0.823,-2.794l-5.05,-9.848c-0.321,-0.638 -0.976,-1.041 -1.69,-1.041c-0.339,0 -0.673,0.091 -0.965,0.264l-14.599,9.231c-0.523,0.354 -0.837,0.946 -0.837,1.577c0,0.418 0.138,0.825 0.392,1.157l6.676,8.813c0.682,0.881 1.946,1.089 2.874,0.471l12.376,-7.83Zm55.584,-15.907c1.109,-0.097 1.966,-1.044 1.951,-2.157l-0.285,-11.056c-0.018,-1.028 -0.869,-1.864 -1.898,-1.864c-0.051,0 -0.102,0.002 -0.153,0.006l-17.194,1.911c-0.929,0.128 -1.628,0.93 -1.628,1.867c0,0.133 0.014,0.266 0.042,0.396l2.19,10.837c0.234,1.089 1.278,1.824 2.382,1.679l14.593,-1.619Zm14.938,-15.071c-0.051,-0.004 -0.102,-0.006 -0.153,-0.006c-1.028,0 -1.879,0.836 -1.898,1.864l-0.278,11.056c-0.015,1.113 0.842,2.06 1.951,2.157l14.599,1.619c1.104,0.145 2.148,-0.59 2.383,-1.679l2.189,-10.837c0.028,-0.13 0.043,-0.263 0.043,-0.396c0,-0.937 -0.7,-1.739 -1.629,-1.867l-17.207,-1.911Z" />
+      </svg>
+    ),
+    aspectRatio: 1760 / 360,
+  },
+  Frontier: {
+    svg: (
+      <svg viewBox="0 0 1660 360">
+        <path d="M45.486,228.837c-3.263,-8.993 -5.639,-18.347 -7.014,-27.936c-7.348,-51.219 14.56,-105.293 56.79,-136.889c24.579,-18.389 55.142,-28.589 85.802,-28.821c0.468,-0.003 0.936,-0.004 1.404,-0.003c66.066,0.321 129.46,48.82 142.186,118.907c9.755,53.724 -13.165,112.106 -59.089,144.575c-39.728,28.09 -93.877,34.286 -139.171,15.455c-25.026,-10.404 -46.702,-28.11 -62.431,-50.185c-2.168,0.807 -4.515,1.247 -6.963,1.247c-11.038,0 -20,-8.961 -20,-20c0,-6.752 3.354,-12.728 8.486,-16.35Zm26.704,29.356c8.654,12.38 19.403,23.393 32.018,32.312c36.911,26.097 87.412,31.908 129.559,14.386c42.835,-17.807 75.15,-58.615 81.863,-105.41c5.647,-39.366 -7.009,-80.997 -34.265,-110.642c-25.02,-27.213 -61.266,-43.367 -98.491,-43.649c-0.437,-0.002 -0.873,-0.002 -1.31,-0.002c-61.569,0.299 -120.543,45.497 -132.379,110.693c-4.202,23.147 -1.898,47.211 6.026,69.386c0.59,-0.053 1.186,-0.08 1.789,-0.08c11.038,0 20,8.962 20,20c0,4.964 -1.812,9.508 -4.81,13.006Zm90.029,-175.973c1.435,-9.633 9.751,-17.033 19.781,-17.033c10.052,0 18.382,7.432 19.79,17.096c29.375,6.19 55.992,25.714 69.806,53.471c13.535,27.195 13.676,60.72 0.403,88.048c-16.135,33.219 -51.891,56.275 -89.677,56.385c-37.363,0.109 -73.819,-22.186 -90.451,-56.655c-18.869,-39.105 -9.555,-89.82 25.383,-119.533c12.806,-10.89 28.443,-18.395 44.965,-21.779Zm38.509,9.996c-2.848,7.575 -10.163,12.971 -18.728,12.971c-8.545,0 -15.845,-5.369 -18.707,-12.916c-28.922,6.421 -54.727,27.246 -65.638,56.47c-9.39,25.149 -6.781,54.456 7.099,77.632c18.504,30.894 56.53,48.962 93.173,42.408c40.212,-7.192 73.644,-43.868 74.068,-87.721c0.225,-23.253 -8.93,-46.44 -25.233,-63.344c-12.351,-12.805 -28.558,-21.719 -46.034,-25.5Zm561.55,164.715c-14.534,0 -27.634,-3.321 -39.3,-9.964c-11.667,-6.642 -20.8,-15.794 -27.4,-27.454c-6.6,-11.66 -9.9,-24.768 -9.9,-39.326c0,-14.557 3.3,-27.666 9.9,-39.326c6.6,-11.66 15.733,-20.811 27.4,-27.454c11.666,-6.642 24.766,-9.964 39.3,-9.964c14.533,0 27.6,3.322 39.2,9.964c11.6,6.643 20.733,15.794 27.4,27.454c6.666,11.66 10,24.769 10,39.326c0,14.558 -3.334,27.666 -10,39.326c-6.667,11.66 -15.8,20.812 -27.4,27.454c-11.6,6.643 -24.667,9.964 -39.2,9.964Zm-272.8,-123.384l-73.4,0l0,39.22l64.8,0l0,27.56l-64.8,0l0,54.06l-32.4,0l0,-148.4l105.8,0l0,27.56Zm745.4,120.84l-32.4,0l0,-148.4l32.4,0l0,148.4Zm157,-120.84l-73.6,0l0,32.224l65,0l0,26.712l-65,0l0,34.344l76.2,0l0,27.56l-108.4,0l0,-148.4l105.8,0l0,27.56Zm-224.8,0.424l-44.8,0l0,120.416l-32.4,0l0,-120.416l-44.8,0l0,-27.984l122,0l0,27.984Zm-605.2,120.416l-32.4,0l0,-148.4l60.6,0c12.4,0 23.166,2.191 32.3,6.572c9.133,4.382 16.166,10.6 21.1,18.656c4.933,8.056 7.4,17.596 7.4,28.62c0,11.024 -2.5,20.529 -7.5,28.514c-5,7.986 -12.1,14.098 -21.3,18.338l31.4,47.7l-34.8,0l-27,-41.34l-29.8,0l0,41.34Zm908.8,0l-32.4,0l0,-148.4l60.6,0c12.4,0 23.166,2.191 32.3,6.572c9.133,4.382 16.166,10.6 21.1,18.656c4.933,8.056 7.4,17.596 7.4,28.62c0,11.024 -2.5,20.529 -7.5,28.514c-5,7.986 -12.1,14.098 -21.3,18.338l31.4,47.7l-34.8,0l-27,-41.34l-29.8,0l0,41.34Zm-557.4,0l-32,0l0,-148.4l26.8,0l69.6,90.1l0,-90.1l32,0l0,148.4l-26.6,0l-69.8,-90.1l0,90.1Zm-685.29,-55.422c-3.502,8.758 -9.368,16.598 -17.292,22.163c-18.438,12.951 -45.747,11.656 -62.967,-4.545c-19.099,-17.967 -20.825,-51.517 -1.5,-71.317c9.208,-9.436 22.338,-14.951 35.449,-15.078c0.215,0 0.43,0 0.645,0c18.862,0.183 37.106,11.752 45.127,29.148c9.885,1.21 17.55,9.643 17.55,19.851c0,10.023 -7.389,18.334 -17.012,19.778Zm534.29,28.71c8.266,0 15.733,-2.014 22.4,-6.042c6.666,-4.028 11.9,-9.646 15.7,-16.854c3.8,-7.208 5.7,-15.405 5.7,-24.592c0,-9.186 -1.9,-17.384 -5.7,-24.592c-3.8,-7.208 -9.034,-12.826 -15.7,-16.854c-6.667,-4.028 -14.134,-6.042 -22.4,-6.042c-8.267,0 -15.734,2.014 -22.4,6.042c-6.667,4.028 -11.9,9.646 -15.7,16.854c-3.8,7.208 -5.7,15.406 -5.7,24.592c0,9.187 1.9,17.384 5.7,24.592c3.8,7.208 9.033,12.826 15.7,16.854c6.666,4.028 14.133,6.042 22.4,6.042Zm-545.321,-66.802c-1.666,-3.063 -3.733,-5.916 -6.176,-8.452c-7.299,-7.573 -17.565,-12.064 -28.262,-12.23c-0.26,-0.003 -0.519,-0.003 -0.778,-0.003c-20.197,0.195 -39.047,17.113 -39.72,38.705c-0.725,23.259 21.885,45.336 47.306,40.624c12.156,-2.252 22.937,-10.634 28.513,-21.652c-7.506,-2.882 -12.84,-10.162 -12.84,-18.678c0,-8.177 4.919,-15.215 11.957,-18.314Zm371.321,-26.902l-26.4,0l0,51.728l26.4,0c9.733,0 17.133,-2.261 22.2,-6.784c5.066,-4.522 7.6,-10.882 7.6,-19.08c0,-8.338 -2.534,-14.734 -7.6,-19.186c-5.067,-4.452 -12.467,-6.678 -22.2,-6.678Zm908.8,0l-26.4,0l0,51.728l26.4,0c9.733,0 17.133,-2.261 22.2,-6.784c5.066,-4.522 7.6,-10.882 7.6,-19.08c0,-8.338 -2.534,-14.734 -7.6,-19.186c-5.067,-4.452 -12.467,-6.678 -22.2,-6.678Z" />
+      </svg>
+    ),
+    aspectRatio: 1660 / 360,
+  },
+  Labs: {
+    svg: (
+      <svg viewBox="0 0 1080 360">
+        <path d="M180.848,67.303l-0.889,-0.519c-5.51,-3.345 -7.718,-10.987 -4.513,-16.854c2.706,-4.703 5.067,-9.638 8.18,-14.07c2.345,-3.341 6.048,-5.345 10.406,-5.308c2.071,0.04 4.107,0.586 5.924,1.592c34.536,19.506 68.801,39.49 103.05,59.495c5.528,3.356 7.716,10.989 4.511,16.856c-4.194,7.273 -7.341,15.914 -13.753,18.479c-3.451,1.38 -7.466,1.118 -10.756,-0.695l-1.16,-0.658l-17.125,29.661c18.787,34.685 17.592,79.466 -4.029,113.383c-24.918,39.088 -75.505,59.997 -121.975,47.997c-39.59,-10.223 -72.186,-43.332 -81.156,-84.419c-1.761,-8.068 -2.555,-16.324 -2.432,-24.576c-0.12,-0.57 -0.118,-1.176 0.042,-1.807c0.893,-28.668 12.817,-57.124 33.668,-77.365c19.805,-19.227 46.728,-30.655 74.525,-31.516l0.354,-0.01l17.128,-29.666Zm87.216,144.691c-14.204,-1.23 -27.343,-9.541 -40.909,-12.047c-12.529,-2.314 -25.048,2.101 -37.835,6.971c-14.258,5.395 -29.778,6.768 -44.404,1.749c-14.524,-4.984 -28.314,-12.568 -45.604,-7.57c-11.5,3.324 -22.293,9.438 -34.187,10.786c0.255,8.242 1.502,16.465 3.766,24.444c11.21,39.508 48.05,70.832 90.999,73.645c40.52,2.655 81.519,-20.94 99.031,-58.861c5.672,-12.284 8.718,-25.645 9.143,-39.117Zm643.814,44.937c-11.067,0 -21.767,-1.59 -32.1,-4.77c-10.334,-3.18 -18.634,-7.314 -24.9,-12.402l11,-25.864c6,4.664 13.133,8.41 21.4,11.236c8.266,2.827 16.533,4.24 24.8,4.24c9.2,0 16,-1.448 20.4,-4.346c4.4,-2.897 6.6,-6.748 6.6,-11.554c0,-3.533 -1.3,-6.466 -3.9,-8.798c-2.6,-2.332 -5.934,-4.204 -10,-5.618c-4.067,-1.413 -9.567,-2.968 -16.5,-4.664c-10.667,-2.685 -19.4,-5.37 -26.2,-8.056c-6.8,-2.685 -12.634,-6.996 -17.5,-12.932c-4.867,-5.936 -7.3,-13.85 -7.3,-23.744c0,-8.621 2.2,-16.43 6.6,-23.426c4.4,-6.996 11.033,-12.543 19.9,-16.642c8.866,-4.098 19.7,-6.148 32.5,-6.148c8.933,0 17.666,1.131 26.2,3.392c8.533,2.262 16,5.512 22.4,9.752l-10,26.076c-12.934,-7.773 -25.867,-11.66 -38.8,-11.66c-9.067,0 -15.767,1.555 -20.1,4.664c-4.334,3.11 -6.5,7.208 -6.5,12.296c0,5.088 2.5,8.869 7.5,11.342c5,2.474 12.633,4.912 22.9,7.314c10.666,2.686 19.4,5.371 26.2,8.056c6.8,2.686 12.633,6.926 17.5,12.72c4.866,5.795 7.3,13.639 7.3,23.532c0,8.48 -2.234,16.218 -6.7,23.214c-4.467,6.996 -11.167,12.544 -20.1,16.642c-8.934,4.099 -19.8,6.148 -32.6,6.148Zm-425.6,-2.544l-102.6,0l0,-148.4l32.4,0l0,120.416l70.2,0l0,27.984Zm281,0l-72.4,0l0,-148.4l68.4,0c17.066,0 30.166,3.463 39.3,10.388c9.133,6.926 13.7,16.324 13.7,28.196c0,7.208 -1.634,13.639 -4.9,19.292c-3.267,5.654 -7.834,10.106 -13.7,13.356c8,2.686 14.266,7.173 18.8,13.462c4.533,6.29 6.8,14.028 6.8,23.214c0,13.003 -4.767,23.002 -14.3,29.998c-9.534,6.996 -23.433,10.494 -41.7,10.494Zm-105,0l-34,0l-12.4,-31.8l-65,0l-12.4,31.8l-33.2,0l62.4,-148.4l32,0l62.6,148.4Zm102.6,-25.864c8.533,0 14.966,-1.484 19.3,-4.452c4.333,-2.968 6.5,-7.561 6.5,-13.78c0,-12.296 -8.6,-18.444 -25.8,-18.444l-37.8,0l0,36.676l37.8,0Zm-575.394,-156.178l-18.525,32.086c-2.116,3.167 -3.565,2.455 -6.878,2.532c-35.878,0.97 -70.33,22.19 -87.211,53.934c-6.717,12.631 -10.618,26.671 -11.566,40.934c0.191,-0.022 0.38,-0.046 0.569,-0.073c12.529,-1.783 23.611,-9.19 36.271,-11.626c10.288,-1.979 21.02,-0.569 30.904,3.122c7.397,2.763 14.536,6.343 22.197,7.931c19.443,4.03 38.605,-9.592 57.144,-11.693c2.158,-0.244 4.33,-0.353 6.5,-0.345c11.312,0.09 22.223,4.377 32.665,8.422c5.281,2.037 10.753,3.781 16.341,4.369c-1.003,-15.381 -5.425,-30.594 -13.257,-44.198c0,0 -0.944,-2.124 -0.496,-3.791c0.151,-0.563 0.243,-0.685 0.499,-1.204l18.514,-32.067c-27.987,-15.942 -55.85,-32.102 -83.671,-48.333Zm416.194,124.166l-22.2,-56.816l-22.2,56.816l44.4,0Zm121.4,-64.66l0,34.98l32.2,0c8,0 14.066,-1.484 18.2,-4.452c4.133,-2.968 6.2,-7.349 6.2,-13.144c0,-5.794 -2.067,-10.14 -6.2,-13.038c-4.134,-2.897 -10.2,-4.346 -18.2,-4.346l-32.2,0Zm-533.243,-91.3c-0.861,0.015 -1.686,0.486 -2.14,1.234c-2.562,4.342 -5.183,8.653 -7.56,13.1c-0.592,1.129 -0.169,2.62 0.913,3.29c33.857,20.556 68.045,40.609 102.907,59.412c1.12,0.581 2.642,0.161 3.304,-0.932c2.541,-4.375 5.237,-8.665 7.586,-13.145c0.571,-1.138 0.138,-2.614 -0.956,-3.279c-34.192,-19.972 -67.994,-40.66 -102.879,-59.395c-0.328,-0.171 -0.685,-0.266 -1.055,-0.283c-0.04,-0.001 -0.08,-0.002 -0.12,-0.002Z" />
+      </svg>
+    ),
+    aspectRatio: 1080 / 360,
+  },
+  OperationalUI: {
+    svg: (
+      <svg viewBox="0 0 2170 361">
+        <path d="M254.202,56.201c26.119,15.671 47.637,39.34 59.734,68.377c18.488,44.378 13.211,97.876 -14.703,138.103c-26.549,38.26 -72.302,62.507 -119.233,62.507c-37.011,0 -73.619,-14.739 -100.336,-40.359c-23.335,-22.378 -38.941,-52.768 -43.319,-84.833c-5.917,-43.332 9.404,-89.114 38.714,-119.832c26.723,-28.005 64.651,-44.675 104.005,-44.973c0.468,-0.003 0.936,-0.004 1.404,-0.003c24.436,0.119 48.553,6.827 69.861,18.766c2.252,0.202 2.445,0.73 3.104,1.351c0.293,0.278 0.551,0.577 0.769,0.896Zm-2.475,9.752l-22.396,38.791c-0.031,0.051 -0.063,0.1 -0.096,0.149c3.528,2.292 6.894,4.833 10.06,7.608c30.686,26.905 40.147,75.552 18.635,112.709c-17.951,31.006 -55.492,49.708 -92.183,43.853c-9.167,-1.463 -18.051,-4.419 -26.309,-8.642c-0.035,0.07 -0.071,0.14 -0.109,0.21l-22.427,38.845c34.905,18.236 78.023,20.732 114.865,5.415c43.072,-17.906 75.595,-59.481 81.98,-106.248c5.362,-39.27 -7.429,-80.488 -34.382,-109.804c-8.141,-8.855 -17.472,-16.539 -27.638,-22.886Zm-8.67,-4.983c-19.032,-10.036 -40.44,-15.615 -62.183,-15.78c-0.437,-0.002 -0.873,-0.002 -1.31,-0.002c-61.569,0.299 -120.539,45.495 -132.379,110.694c-9.085,50.028 12.248,104.364 55.024,134.623c1.967,1.392 3.973,2.726 6.015,4.002l22.445,-38.876c0.047,-0.077 0.097,-0.152 0.148,-0.226c-13.513,-8.877 -24.759,-21.343 -31.932,-36.207c-15.789,-32.718 -9.622,-74.693 16.538,-101.675c16.512,-17.031 40.054,-27.056 63.705,-27.331c0.388,-0.003 0.775,-0.005 1.163,-0.004c13.941,0.067 27.835,3.481 40.285,9.739c0.031,-0.061 0.062,-0.122 0.095,-0.183l22.386,-38.774Zm-63.315,39.218c-44.396,0.215 -86.012,41.031 -78.966,91.184c4.539,32.305 30.673,60.657 63.349,67.238c27.503,5.541 57.494,-4.29 76.356,-26.082c20.865,-24.104 25.71,-60.611 10.203,-89.819c-13.498,-25.423 -41.388,-42.381 -70.426,-42.521c-0.172,0 -0.344,0 -0.516,0Zm271.736,156.743c-14.534,0 -27.634,-3.321 -39.3,-9.964c-11.667,-6.642 -20.8,-15.794 -27.4,-27.454c-6.6,-11.66 -9.9,-24.768 -9.9,-39.326c0,-14.557 3.3,-27.666 9.9,-39.326c6.6,-11.66 15.733,-20.811 27.4,-27.454c11.666,-6.642 24.766,-9.964 39.3,-9.964c14.533,0 27.6,3.322 39.2,9.964c11.6,6.643 20.733,15.794 27.4,27.454c6.666,11.66 10,24.769 10,39.326c0,14.558 -3.334,27.666 -10,39.326c-6.667,11.66 -15.8,20.812 -27.4,27.454c-11.6,6.643 -24.667,9.964 -39.2,9.964Zm1055.8,0c-14.534,0 -27.634,-3.321 -39.3,-9.964c-11.667,-6.642 -20.8,-15.794 -27.4,-27.454c-6.6,-11.66 -9.9,-24.768 -9.9,-39.326c0,-14.557 3.3,-27.666 9.9,-39.326c6.6,-11.66 15.733,-20.811 27.4,-27.454c11.666,-6.642 24.766,-9.964 39.3,-9.964c14.533,0 27.6,3.322 39.2,9.964c11.6,6.643 20.733,15.794 27.4,27.454c6.666,11.66 10,24.769 10,39.326c0,14.558 -3.334,27.666 -10,39.326c-6.667,11.66 -15.8,20.812 -27.4,27.454c-11.6,6.643 -24.667,9.964 -39.2,9.964Zm-119,-2.544l-32.4,0l0,-148.4l32.4,0l0,148.4Zm621.2,-27.984l70.2,0l0,27.984l-102.6,0l0,-148.4l32.4,0l0,120.416Zm-1170.8,-92.856l-73.6,0l0,32.224l65,0l0,26.712l-65,0l0,34.344l76.2,0l0,27.56l-108.4,0l0,-148.4l105.8,0l0,27.56Zm-235.8,120.84l-32.4,0l0,-148.4l60.6,0c12.4,0 23.166,2.191 32.3,6.572c9.133,4.382 16.166,10.6 21.1,18.656c4.933,8.056 7.4,17.596 7.4,28.62c0,10.883 -2.467,20.388 -7.4,28.514c-4.934,8.127 -11.967,14.346 -21.1,18.656c-9.133,4.311 -19.9,6.466 -32.3,6.466l-28.2,0l0,40.916Zm717.6,-120.416l-44.8,0l0,120.416l-32.4,0l0,-120.416l-44.8,0l0,-27.984l122,0l0,27.984Zm-403,120.416l-32.4,0l0,-148.4l60.6,0c12.4,0 23.166,2.191 32.3,6.572c9.133,4.382 16.166,10.6 21.1,18.656c4.933,8.056 7.4,17.596 7.4,28.62c0,11.024 -2.5,20.529 -7.5,28.514c-5,7.986 -12.1,14.098 -21.3,18.338l31.4,47.7l-34.8,0l-27,-41.34l-29.8,0l0,41.34Zm740.8,0l-32,0l0,-148.4l26.8,0l69.6,90.1l0,-90.1l32,0l0,148.4l-26.6,0l-69.8,-90.1l0,90.1Zm286.2,0l-34,0l-12.4,-31.8l-65,0l-12.4,31.8l-33.2,0l62.4,-148.4l32,0l62.6,148.4Zm-755.8,0l-34,0l-12.4,-31.8l-65,0l-12.4,31.8l-33.2,0l62.4,-148.4l32,0l62.6,148.4Zm-737.2,-26.712c8.266,0 15.733,-2.014 22.4,-6.042c6.666,-4.028 11.9,-9.646 15.7,-16.854c3.8,-7.208 5.7,-15.405 5.7,-24.592c0,-9.186 -1.9,-17.384 -5.7,-24.592c-3.8,-7.208 -9.034,-12.826 -15.7,-16.854c-6.667,-4.028 -14.134,-6.042 -22.4,-6.042c-8.267,0 -15.734,2.014 -22.4,6.042c-6.667,4.028 -11.9,9.646 -15.7,16.854c-3.8,7.208 -5.7,15.406 -5.7,24.592c0,9.187 1.9,17.384 5.7,24.592c3.8,7.208 9.033,12.826 15.7,16.854c6.666,4.028 14.133,6.042 22.4,6.042Zm1055.8,0c8.266,0 15.733,-2.014 22.4,-6.042c6.666,-4.028 11.9,-9.646 15.7,-16.854c3.8,-7.208 5.7,-15.405 5.7,-24.592c0,-9.186 -1.9,-17.384 -5.7,-24.592c-3.8,-7.208 -9.034,-12.826 -15.7,-16.854c-6.667,-4.028 -14.134,-6.042 -22.4,-6.042c-8.267,0 -15.734,2.014 -22.4,6.042c-6.667,4.028 -11.9,9.646 -15.7,16.854c-3.8,7.208 -5.7,15.406 -5.7,24.592c0,9.187 1.9,17.384 5.7,24.592c3.8,7.208 9.033,12.826 15.7,16.854c6.666,4.028 14.133,6.042 22.4,6.042Zm380.6,-31.164l-22.2,-56.816l-22.2,56.816l44.4,0Zm-755.8,0l-22.2,-56.816l-22.2,56.816l44.4,0Zm-188.2,-62.54l-26.4,0l0,51.728l26.4,0c9.733,0 17.133,-2.261 22.2,-6.784c5.066,-4.522 7.6,-10.882 7.6,-19.08c0,-8.338 -2.534,-14.734 -7.6,-19.186c-5.067,-4.452 -12.467,-6.678 -22.2,-6.678Zm-314.6,0l-26.4,0l0,51.516l26.4,0c9.733,0 17.133,-2.226 22.2,-6.678c5.066,-4.452 7.6,-10.776 7.6,-18.974c0,-8.338 -2.534,-14.734 -7.6,-19.186c-5.067,-4.452 -12.467,-6.678 -22.2,-6.678Z" />
+      </svg>
+    ),
+    aspectRatio: 2160 / 360,
+  },
+}
