+++ packages/components/src/index.ts
@@ -18,7 +18,7 @@ export { default as Chip } from "./Chip/Chip"
 export { default as Code } from "./Code/Code"
 export { default as Confirm } from "./Confirm/Confirm"
 export { default as ContextMenu } from "./ContextMenu/ContextMenu"
-export { default as ContiamoLogo } from "./ContiamoLogo/ContiamoLogo"
+export { default as Logo } from "./Logo/Logo"
 export { default as Grid } from "./Grid/Grid"
 export { default as HeaderMenu } from "./HeaderMenu/HeaderMenu"
 export { default as Hint } from "./Hint/Hint"
